+++ src/lib.rs
@@ -274,6 +274,7 @@ impl AppBuilder {
                 secrets: self.secrets_settings,
                 readyz_require_all_sources: self.readyz_require_all_sources,
                 metrics_require_auth: self.metrics_require_auth,
+                cors_allowed_origins: self.cors_allowed_origins,
                 shutdown_grace_seconds: self.shutdown_grace_seconds,
                 refresh_timeout_seconds: self.refresh_timeout_seconds,
                 refresh_max_concurrent: self.refresh_max_concurrent,
@@ -300,7 +301,6 @@ impl AppBuilder {
             Arc::new(adapters::r#in::http::auth::ApiKeyRegistry::new(
                 self.api_keys,
             )),
-            self.cors_allowed_origins,
             self.max_body_bytes,
         );
         (router, state)